Removed usage of level.time, replaced with level.framenum and respectively edited the functions that use it.

+++ trunk/game/cc_entitytypes.h
@@ -61,7 +61,7 @@ public:
 class CThinkableEntity abstract : public virtual CBaseEntity
 {
 public:
-	float		NextThink;
+	int32		NextThink;
 
 	CThinkableEntity ();
 	CThinkableEntity (int index);
